Verification and clean up
I followed the instructions on another computer and verified that they worked. There were some slight variations which I decided to favor since the computer was new to OpenGL (as opposed to the other which already had some tinkering done).

I also made a few mainly cosmetic changes to the Readme and to the code to remove warnings and to report errors (though there shouldn't be any).